add controller and update useCase

+++ application/usecases/accounts/DeleteAccountUseCase.ts
@@ -4,12 +4,18 @@ import { AccountRepositoryInterface } from "../../repositories/AccountRepository
 export class DeleteAccountUseCase {
     public constructor ( private accountRepository: AccountRepositoryInterface){}
     
-    public async execute(accountNumber: number) {
+    public async execute(accountNumber: number): Promise<void | Error> {
+        const existingAccount = await this.accountRepository.getOneAccountByAccountNumber(accountNumber)
+        
+        if(existingAccount instanceof Error) {
+            return existingAccount;
+        }
 
         const account = await this.accountRepository.deleteAccount(accountNumber);
 
-        if(account instanceof AccountNotFoundError) {
+        if(account instanceof Error) {
             return account;
         }
+
     }
 }
